Medicinaai/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "openai>=1.63.2",
    "flask-mongoengine>=1.0.0",
    "scikit-learn>=1.6.1",
    "numpy>=2.2.3",
    "werkzeug>=3.1.3",
    "ollama>=0.4.7",
    "sqlalchemy>=2.0.38",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.3",
]
